-- Correção de partes do código (controlador de dados de saude comentado p/ reestruturação) -- Adicionado funcionalidade de lembretes

+++ www/js/controllers.js
@@ -118,9 +118,8 @@ medApp.controllers = {
         page.querySelector('#crm-perfil').innerHTML = data[0].CRM;
         page.querySelector('#esp-perfil').innerHTML = data[0].especialidade;
         page.querySelector('#tel-perfil').innerHTML = data[0].telefone;
+        page.querySelector('#email-perfil').innerHTML = data[0].email;
       });
-
-      page.querySelector('#email-perfil').innerHTML = Math.random();
 
     });
 
@@ -132,16 +131,25 @@ medApp.controllers = {
         {nome: page.querySelector('.profile-name').innerHTML,
          CRM: page.querySelector('#crm-perfil').innerHTML,
          esp: page.querySelector('#esp-perfil').innerHTML,
-         tel: page.querySelector('#tel-perfil').innerHTML
-                }});
+         tel: page.querySelector('#tel-perfil').innerHTML,
+         email: page.querySelector('#email-perfil').innerHTML
+        }});
 
     };
 
     // Realiza o logoff
     page.querySelector('#logoff').onclick = function() {
 
-      medApp.services.deleteIdMedico();
-      document.querySelector('#loginNav').resetToPage( 'login.html', {options: {animation: 'fade'}});
+      ons.notification.confirm({message: 'Tem certeza?'})
+        .then( function(confirm){
+
+          if(confirm) {
+            medApp.services.deleteIdMedico();
+            document.querySelector('#loginNav').resetToPage( 'login.html', {options: {animation: 'fade'}});
+          };
+
+        });
+
     };
 
     // Realiza a atualização do perfil com pull
@@ -167,7 +175,6 @@ medApp.controllers = {
 
     pullHook.onAction = function(done) {
       setTimeout(done, 1000);
-      page.querySelector('#email-perfil').innerHTML = Math.random();
     };
 
   },
@@ -219,6 +226,14 @@ medApp.controllers = {
     page.querySelector('.profile-image').src = page.data.img;
     medApp.services.setIdPaciente($('#idPaciente')); //TODO --> ver se prontuário é retornado e faz papel de ID.
 
+    // Chama página de edição de dados do paciente
+    page.querySelector('#pacienteeditar').onclick = function() {
+
+      document.querySelector('#pacienteNav').pushPage('html/editarpaciente.html');
+
+    };
+
+    /*
     // Chama página de dados de saúde
     page.querySelector('#graf1').onclick = function() {
 
@@ -233,6 +248,8 @@ medApp.controllers = {
           console.log('Os dados retornados são: ' + medApp.services.getDadosEstaticos());
           });
       });
+
+    },
 
       /////////////////////////////////////
       ///Controle dos Gráficos de saúde ///
@@ -391,15 +408,9 @@ medApp.controllers = {
 
 
 
-
+      
     };
-
-    // Chama página de edição de dados do paciente
-    page.querySelector('#pacienteeditar').onclick = function() {
-
-      document.querySelector('#pacienteNav').pushPage('html/editarpaciente.html');
-    };
-
+    */
   },
 
   /////////////////////////////////////
@@ -414,7 +425,8 @@ medApp.controllers = {
       nomeEdit: page.data.nome,
       crmEdit: page.data.CRM,
       espEdit: page.data.esp, 
-      telEdit: page.data.tel
+      telEdit: page.data.tel,
+      emailEdit: page.data.email,
 
     };
 
@@ -422,6 +434,7 @@ medApp.controllers = {
     $('#crm-medico').val(dadosEdit.crmEdit);
     $('#esp-medico').val(dadosEdit.espEdit);
     $('#tel-medico').val(dadosEdit.telEdit);
+    $('#email-medico').val(dadosEdit.emailEdit);
 
     // Botão salvar altera os dados no servidor se houve mudanças 
     page.querySelector('#salvar-med').onclick = function() {
@@ -431,7 +444,8 @@ medApp.controllers = {
         nomeEdit: $('#nome-medico').val(),
         crmEdit: $('#crm-medico').val(),
         espEdit: $('#esp-medico').val(), 
-        telEdit: $('#tel-medico').val()
+        telEdit: $('#tel-medico').val(),
+        emailEdit: $('#email-medico').val()
 
       };
 
@@ -474,7 +488,7 @@ medApp.controllers = {
 
     $('#prontuario-pac').val(dadosEdit.prontEdit);
 
- 	$('#idade-pac').val(dadosEdit.idadeEdit);
+ 	  $('#idade-pac').val(dadosEdit.idadeEdit);
 
   	$('#email-pac').val(dadosEdit.emailEdit);
 
@@ -488,7 +502,7 @@ medApp.controllers = {
         obsEdit: $('#obs-pac').val(),
         prontEdit: $('#pront-pac').val(),
         idadeEdit: $('#idade-pac').val(),
-   		emailEdit: $('#email-pac').val(),
+   		  emailEdit: $('#email-pac').val(),
       };
 
       if (medApp.services.checkEdit(novoEdit, dadosEdit)) {
@@ -513,7 +527,22 @@ medApp.controllers = {
 
     page.querySelector('#add-lembrete').onclick = function() {
 
-      medApp.services.createLembrete();
+      ons.notification.prompt({message: 'Escreva abaixo o lembrete:'})
+        .then(function(texto){
+
+          if( texto === '' ) {
+
+            ons.notification.alert('Insira algum texto!');
+
+          } else {
+
+            medApp.services.createLembrete(texto);
+
+          };
+
+        });
+      
+
     };
   }
 };
